Module NXT.Find

List files on the brick matching a given pattern.

type 'a iterator

An iterator to allow to enumerate files on the brick.

val patt : 'a conn -> string -> 'a iterator Lwt.t

Find.patt conn fpatt returns an iterator listing the filenames mathing the pattern fpatt. The following types of wildcards are accepted:

filename.extension

*.[file type name]

filename.*

*.*

raises File_not_found
if no file was found

val current : 'a iterator -> string Lwt.t

Find.current i returns the current filename.

val current_size : 'a iterator -> int Lwt.t

Find.current_size i returns the current filename size (number of bytes).

val next : 'a iterator -> unit Lwt.t

Execute a new request to the brick to retrieve the next filename matching the pattern.

raises File_not_found
if no more file was found. When this exception is raised, the iterator is closed.

val close : 'a iterator -> unit Lwt.t

close_iterator i closes the iterator i. Closing an already closed iterator does nothing.

val iter : 'a conn -> f:(string -> int -> unit Lwt.t) -> string -> unit Lwt.t

iter f fpatt iterates f name size on all the filenames matching the pattern fpatt (see Mindstorm_lwt.NXT.Find.patt for the accepted patterns).

val map : 'a conn -> f:(string -> int -> 'b Lwt.t) -> string -> 'b list Lwt.t

map f fpatt maps f name size on all the filenames matching the pattern fpatt and return the list formed of those. (See Mindstorm_lwt.NXT.Find.patt for the accepted patterns.)

val fold : 'a conn -> f:(string -> int -> 'b -> 'b Lwt.t) -> string -> 'b -> 'b Lwt.t

fold f fpatt a0 folds f on all the filenames matching the pattern fpatt (see Mindstorm_lwt.NXT.Find.patt for the accepted patterns).